backend/requirements.txt
python-dotenv==1.0.0
google-genai>=1.0.0
chromadb==0.4.15
sentence-transformers>=5.0.0
pypdf==3.17.4
requests>=2.31.0
fastapi
uvicorn
boto3==1.26.137
apscheduler>=3.10.0
websockets>=13.0.0,<15.1.0
h11>=0.13,<0.15